renamed image processing to registration
